[sw] Add simple XADC driver
Adds a simple (X)ADC driver which exposes the ability to read the
relevant ADC status registers that are changed by ADC sampling,
and to set the clock divider and power down functionality. This
represents a minimal set of XADC functionality that can then be
expanded on in the future.