Vendor in automotive demo library
The automotive demo runs on both CHERIoT and non-CHERIoT, and to support
the build flow for this, it is implemented in both the `sonata-software`
and `sonata-system repositories. To avoid code duplication, the demo's
library code (for which implementations just have to provide callbacks
to initialise and then call relevant functions to carry out the demo) is
stored in a separate git repository and then vendored into this repo, so
that it can be used.