Add LCD initialisation/wrapper for automotive demo
Adds a simple LCD driver/wrapper to be used in the automotive demo's
sending firmware for non-CHERIoT systems. This is based on the existing
code in the `lcd_st7735` demo that is already using the LCD, but is
stripped down to the bare essentials of just initialising the LCD for
use and removing some of the initialised defaults. This then allows for
LCD callbacks to be implemented for use by the automotive demo library,
which can then use the LCD when running in baremetal non-CHERI.