Add Ethernet wrapper/driver for automotive demo
The automotive demo requires the ability to send ethernet frames
(between two Sonata boards), but does not want to use an entire TCP/IP
stack on top of it; data is being sent at the data-link layer level.

This commit adds a simple wrapper/driver for using the Ethernet in this
way for the automotive demo, which is based on the existing code already
present in the legacy ethernet demo. It simply takes the existing code
and extracts the LWIP TCP/IP interface, to allow the functions to use a
simplified network interface (containing just the SPI handle, MAC source
and maximum transfer units) and a simplified frame buffer (storing just
the payload and length). This then enables ethernet to be used for the
demo without having to define TCP/IP-level logic in all parts of the
demo.