Add sending (board 2) auto demo legacy firmware
Implements the base sending firmware (board 1) for the automotive demo,
running in legacy mode (non-CHERI), baremetal. Implements the common
automotive demo library, passing in callbacks with appropriate driver /
system functionality so that the automotive demo functions on the
sending board side. This commit does not include support for the
analogue pedal demo, which requires use of the ADC.

The sending board is responsible for displaying a menu allowing the user
to select a demo application, and then either simulating a pedal or
polling a pedal (depending on the app selected), and transmitting this
pedal information to the receiving board (board 2) via an Ethernet
connection. It is also responsible for simultaneously running another
separate task that is bugged, such that it accidentally overwrites the
acceleration pedal value at some point. Linker script sections are used
to place these tasks contiguously in memory, so that the out-of-bounds
access is guaranteed to overwrite the acceleration, and not cause any
other type of undefined behaviour.